webapi/webapi-webspeech-w3c-tests/webspeech/blink/speechrecognition-basics.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/speech/scripted/basics.html
-->

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Speech API Test: the basics of the Speech JavaScript API</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

setup(function () {
  window.speechReco = new webkitSpeechRecognition();
});

test(function () {
  assert_true('webkitSpeechRecognition' in self);
}, "Test SpeechRecognition existence");

test(function () {
  assert_not_equals(webkitSpeechRecognition, null);
}, "Test SpeechRecognition is not null");

test(function () {
  assert_not_equals(speechReco, null);
}, "Test new SpeechRecognition is not null");

test(function () {
  assert_true('grammars' in speechReco);
}, "Test grammars existence");

test(function () {
  assert_not_equals(speechReco.grammars, null);
}, "Test grammars is not null");

test(function () {
  assert_equals(speechReco.grammars.length, 0);
}, "Test length of grammars");

test(function () {
  assert_true('lang' in speechReco);
}, "Test lang existence");

test(function () {
  assert_true('continuous' in speechReco);
}, "Test continuous existence");

test(function () {
  assert_true('interimResults' in speechReco);
}, "Test interimResults existence");

test(function () {
  assert_true('maxAlternatives' in speechReco);
}, "Test maxAlternatives existence");

test(function () {
  assert_true('start' in speechReco);
}, "Test start existence");

test(function () {
  assert_true('stop' in speechReco);
}, "Test stop existence");

test(function () {
  assert_true('abort' in speechReco);
}, "Test abort existence");

test(function () {
  assert_true('onaudiostart' in speechReco);
}, "Test onaudiostart existence");

test(function () {
  assert_true('onsoundstart' in speechReco);
}, "Test onsoundstart existence");

test(function () {
  assert_true('onspeechstart' in speechReco);
}, "Test onspeechstart existence");

test(function () {
  assert_true('onspeechend' in speechReco);
}, "Test onspeechend existence");

test(function () {
  assert_true('onsoundend' in speechReco);
}, "Test onsoundend existence");

test(function () {
  assert_true('onaudioend' in speechReco);
}, "Test onaudioend existence");

test(function () {
  assert_true('onresult' in speechReco);
}, "Test onresult existence");

test(function () {
  assert_true('onnomatch' in speechReco);
}, "Test onnomatch existence");

test(function () {
  assert_true('onerror' in speechReco);
}, "Test onerror existence");

test(function () {
  assert_true('onstart' in speechReco);
}, "Test onstart existence");

test(function () {
  assert_true('onend' in speechReco);
}, "Test onend existence");

test(function () {
  assert_true('addEventListener' in speechReco);
}, "Test addEventListener existence");

test(function () {
  assert_true('removeEventListener' in speechReco);
}, "Test removeEventListener existence");

test(function () {
  assert_true('dispatchEvent' in speechReco);
}, "Test dispatchEvent existence");

test(function () {
  assert_equals(speechReco.lang, "");
}, "Test default value of lang");

test(function () {
  assert_equals(speechReco.continuous, false);
}, "Test default value of continuous");

test(function () {
  assert_equals(speechReco.maxAlternatives, 1);
}, "Test default value of maxAlternatives");

</script>
